requirements.txt
pydantic>=2.6
fastapi>=0.110
uvicorn[standard]>=0.27
streamlit>=1.32
requests>=2.31
trafilatura>=1.8
beautifulsoup4>=4.12
lxml>=5.1
yt-dlp>=2024.4.9
youtube-transcript-api>=1.2.4,<2.0
biopython>=1.83
keybert>=0.8
sentence-transformers>=2.5
yake>=0.4.8
langdetect>=1.0.9
tiktoken>=0.6
pycountry>=23.12
python-dotenv>=1.0
pytest>=8.0
vcrpy>=6.0
